user password select in database and create jsonwebtoken
